refactor(morphology): localize catalog ownership
Move morphology strategy config schemas from the broad domain config facade to op-local or named op-family owners. Rename standard recipe tag catalogs from milestone labels to owner/surface names, document retained shared surfaces, and add categorical guardrails for domain config facade imports and milestone-prefixed recipe tag catalogs.

Refs: openspec/changes/normalize-morphology-catalog-owners
