+++ mods/mod-swooper-maps/src/domain/morphology/ops/mountains-shared/rules/types.ts
@@ -1,6 +1,1 @@
-import type { Static } from "@swooper/mapgen-core/authoring";
-
-import { MountainsConfigSchema } from "../../../config.js";
-
-export type MountainsConfig = Static<typeof MountainsConfigSchema>;
-
+export type { MountainsConfig } from "../config.js";
