+++ mods/mod-swooper-maps/src/domain/morphology/ops/plan-island-chains/contract.ts
@@ -1,10 +1,6 @@
 import { Type, TypedArraySchemas, defineOp } from "@swooper/mapgen-core/authoring";
 
-import { IslandsConfigSchema } from "../../config.js";
-
-const IslandChainsConfigSchema = Type.Object({
-  islands: IslandsConfigSchema,
-});
+import { IslandChainsConfigSchema } from "./config.js";
 
 const IslandEditSchema = Type.Object({
   index: Type.Integer({ minimum: 0, description: "Tile index in row-major order." }),
@@ -23,8 +19,12 @@ const PlanIslandChainsContract = defineOp({
     width: Type.Integer({ minimum: 1, description: "Map width in tiles." }),
     height: Type.Integer({ minimum: 1, description: "Map height in tiles." }),
     landMask: TypedArraySchemas.u8({ description: "Land mask per tile (1=land, 0=water)." }),
-    boundaryCloseness: TypedArraySchemas.u8({ description: "Boundary proximity per tile (0..255)." }),
-    boundaryType: TypedArraySchemas.u8({ description: "Boundary type per tile (1=conv,2=div,3=trans)." }),
+    boundaryCloseness: TypedArraySchemas.u8({
+      description: "Boundary proximity per tile (0..255).",
+    }),
+    boundaryType: TypedArraySchemas.u8({
+      description: "Boundary type per tile (1=conv,2=div,3=trans).",
+    }),
     volcanism: TypedArraySchemas.u8({ description: "Volcanism signal per tile (0..255)." }),
     rngSeed: Type.Integer({ description: "Seed for deterministic island placement." }),
   }),
