+++ mods/mod-swooper-maps/src/domain/morphology/ops/plan-ridges/contract.ts
@@ -1,6 +1,6 @@
 import { Type, TypedArraySchemas, defineOp } from "@swooper/mapgen-core/authoring";
 
-import { MountainsConfigSchema } from "../../config.js";
+import { MountainsConfigSchema } from "../mountains-shared/config.js";
 
 /**
  * Plans ridge (mountain) masks and diagnostic driver surfaces from tectonic belt drivers.
@@ -15,11 +15,19 @@ const PlanRidgesContract = defineOp({
     width: Type.Integer({ minimum: 1, description: "Map width in tiles." }),
     height: Type.Integer({ minimum: 1, description: "Map height in tiles." }),
     landMask: TypedArraySchemas.u8({ description: "Land mask per tile (1=land, 0=water)." }),
-    boundaryCloseness: TypedArraySchemas.u8({ description: "Boundary proximity per tile (0..255)." }),
-    boundaryType: TypedArraySchemas.u8({ description: "Boundary type per tile (BOUNDARY_TYPE values)." }),
+    boundaryCloseness: TypedArraySchemas.u8({
+      description: "Boundary proximity per tile (0..255).",
+    }),
+    boundaryType: TypedArraySchemas.u8({
+      description: "Boundary type per tile (BOUNDARY_TYPE values).",
+    }),
     upliftPotential: TypedArraySchemas.u8({ description: "Uplift potential per tile (0..255)." }),
-    collisionPotential: TypedArraySchemas.u8({ description: "Collision-driven uplift potential per tile (0..255)." }),
-    subductionPotential: TypedArraySchemas.u8({ description: "Subduction-driven uplift potential per tile (0..255)." }),
+    collisionPotential: TypedArraySchemas.u8({
+      description: "Collision-driven uplift potential per tile (0..255).",
+    }),
+    subductionPotential: TypedArraySchemas.u8({
+      description: "Subduction-driven uplift potential per tile (0..255).",
+    }),
     riftPotential: TypedArraySchemas.u8({ description: "Rift potential per tile (0..255)." }),
     tectonicStress: TypedArraySchemas.u8({ description: "Tectonic stress per tile (0..255)." }),
     beltAge: TypedArraySchemas.u8({
@@ -30,7 +38,8 @@ const PlanRidgesContract = defineOp({
   output: Type.Object({
     mountainMask: TypedArraySchemas.u8({ description: "Mask (1/0): mountain tiles." }),
     orogenyPotential: TypedArraySchemas.u8({
-      description: "Orogeny potential per tile (0..255). Diagnostic driver surface (physics-gated).",
+      description:
+        "Orogeny potential per tile (0..255). Diagnostic driver surface (physics-gated).",
     }),
     fracturePotential: TypedArraySchemas.u8({
       description: "Fracture proxy per tile (0..255). Diagnostic driver surface (physics-gated).",
