+++ mods/mod-swooper-maps/src/recipes/standard/stages/map-ecology/steps/plot-effects/contract.ts
@@ -3,13 +3,13 @@ import ecology from "@mapgen/domain/ecology";
 
 import { ecologyArtifacts } from "../../../ecology/artifacts.js";
 import { morphologyArtifacts } from "../../../morphology/artifacts.js";
-import { M4_EFFECT_TAGS } from "../../../../tags.js";
+import { STANDARD_ENGINE_EFFECT_TAGS } from "../../../../tags.js";
 
 const PlotEffectsStepContract = defineStep({
   id: "plot-effects",
   phase: "gameplay",
   requires: [],
-  provides: [M4_EFFECT_TAGS.engine.plotEffectsApplied],
+  provides: [STANDARD_ENGINE_EFFECT_TAGS.engine.plotEffectsApplied],
   artifacts: {
     requires: [morphologyArtifacts.topography, ecologyArtifacts.biomeClassification],
   },
