+++ mods/mod-swooper-maps/src/recipes/standard/stages/map-hydrology/steps/lakes.contract.ts
@@ -1,6 +1,6 @@
 import { Type, defineStep } from "@swooper/mapgen-core/authoring";
 
-import { M10_EFFECT_TAGS } from "../../../tags.js";
+import { MAP_PROJECTION_EFFECT_TAGS } from "../../../tags.js";
 import { mapArtifacts } from "../../../map-artifacts.js";
 import { morphologyArtifacts } from "../../morphology/artifacts.js";
 import { hydrologyHydrographyArtifacts } from "../../hydrology-hydrography/artifacts.js";
@@ -38,7 +38,7 @@ const LakesStepContract = defineStep({
   id: "lakes",
   phase: "gameplay",
   requires: [],
-  provides: [M10_EFFECT_TAGS.map.hydrologyLakesParityCaptured],
+  provides: [MAP_PROJECTION_EFFECT_TAGS.map.hydrologyLakesParityCaptured],
   artifacts: {
     requires: [morphologyArtifacts.topography, hydrologyHydrographyArtifacts.hydrography],
     provides: [
