+++ mods/mod-swooper-maps/src/recipes/standard/stages/map-hydrology/steps/plotRivers.contract.ts
@@ -1,6 +1,6 @@
 import { Type, defineStep } from "@swooper/mapgen-core/authoring";
 
-import { M4_EFFECT_TAGS, M10_EFFECT_TAGS } from "../../../tags.js";
+import { STANDARD_ENGINE_EFFECT_TAGS, MAP_PROJECTION_EFFECT_TAGS } from "../../../tags.js";
 import { mapArtifacts } from "../../../map-artifacts.js";
 import { hydrologyHydrographyArtifacts } from "../../hydrology-hydrography/artifacts.js";
 
@@ -34,8 +34,11 @@ const PlotRiversStepConfigSchema = Type.Object(
 const PlotRiversStepContract = defineStep({
   id: "plot-rivers",
   phase: "gameplay",
-  requires: [M10_EFFECT_TAGS.map.elevationBuilt],
-  provides: [M4_EFFECT_TAGS.engine.riversModeled, M10_EFFECT_TAGS.map.hydrologyRiversParityCaptured],
+  requires: [MAP_PROJECTION_EFFECT_TAGS.map.elevationBuilt],
+  provides: [
+    STANDARD_ENGINE_EFFECT_TAGS.engine.riversModeled,
+    MAP_PROJECTION_EFFECT_TAGS.map.hydrologyRiversParityCaptured,
+  ],
   artifacts: {
     requires: [hydrologyHydrographyArtifacts.hydrography],
     provides: [
