+++ mods/mod-swooper-maps/src/recipes/standard/stages/map-morphology/steps/plotCoasts.contract.ts
@@ -1,13 +1,13 @@
 import { Type, defineStep } from "@swooper/mapgen-core/authoring";
 
-import { M10_EFFECT_TAGS } from "../../../tags.js";
+import { MAP_PROJECTION_EFFECT_TAGS } from "../../../tags.js";
 import { morphologyArtifacts } from "../../morphology/artifacts.js";
 
 const PlotCoastsStepContract = defineStep({
   id: "plot-coasts",
   phase: "gameplay",
   requires: [],
-  provides: [M10_EFFECT_TAGS.map.coastsPlotted],
+  provides: [MAP_PROJECTION_EFFECT_TAGS.map.coastsPlotted],
   artifacts: {
     requires: [morphologyArtifacts.topography, morphologyArtifacts.coastlineMetrics],
     provides: [],
