+++ mods/mod-swooper-maps/src/recipes/standard/stages/map-morphology/steps/plotVolcanoes.contract.ts
@@ -1,13 +1,13 @@
 import { Type, defineStep } from "@swooper/mapgen-core/authoring";
 
-import { M10_EFFECT_TAGS } from "../../../tags.js";
+import { MAP_PROJECTION_EFFECT_TAGS } from "../../../tags.js";
 import { morphologyArtifacts } from "../../morphology/artifacts.js";
 
 const PlotVolcanoesStepContract = defineStep({
   id: "plot-volcanoes",
   phase: "gameplay",
-  requires: [M10_EFFECT_TAGS.map.continentsPlotted],
-  provides: [M10_EFFECT_TAGS.map.volcanoesPlotted],
+  requires: [MAP_PROJECTION_EFFECT_TAGS.map.continentsPlotted],
+  provides: [MAP_PROJECTION_EFFECT_TAGS.map.volcanoesPlotted],
   artifacts: {
     requires: [morphologyArtifacts.topography, morphologyArtifacts.volcanoes],
     provides: [],
