+++ mods/mod-swooper-maps/src/recipes/standard/stages/placement/steps/derive-placement-inputs/contract.ts
@@ -1,7 +1,7 @@
 import { Type, defineStep } from "@swooper/mapgen-core/authoring";
 import placement from "@mapgen/domain/placement";
 
-import { M4_EFFECT_TAGS } from "../../../../tags.js";
+import { STANDARD_ENGINE_EFFECT_TAGS } from "../../../../tags.js";
 import { placementArtifacts } from "../../artifacts.js";
 import { hydrologyHydrographyArtifacts } from "../../../hydrology-hydrography/artifacts.js";
 import { ecologyArtifacts } from "../../../ecology/artifacts.js";
@@ -14,8 +14,8 @@ const DerivePlacementInputsContract = defineStep({
   id: "derive-placement-inputs",
   phase: "placement",
   requires: [
-    M4_EFFECT_TAGS.engine.riversModeled,
-    M4_EFFECT_TAGS.engine.featuresApplied,
+    STANDARD_ENGINE_EFFECT_TAGS.engine.riversModeled,
+    STANDARD_ENGINE_EFFECT_TAGS.engine.featuresApplied,
   ],
   provides: [],
   artifacts: {
