+++ mods/mod-swooper-maps/src/recipes/standard/stages/placement/steps/placement/contract.ts
@@ -1,14 +1,17 @@
 import { Type, defineStep } from "@swooper/mapgen-core/authoring";
 
-import { M4_EFFECT_TAGS, M10_EFFECT_TAGS } from "../../../../tags.js";
+import { STANDARD_ENGINE_EFFECT_TAGS, MAP_PROJECTION_EFFECT_TAGS } from "../../../../tags.js";
 import { placementArtifacts } from "../../artifacts.js";
 import { mapArtifacts } from "../../../../map-artifacts.js";
 
 const PlacementStepContract = defineStep({
   id: "placement",
   phase: "placement",
-  requires: [M10_EFFECT_TAGS.map.landmassRegionsPlotted],
-  provides: [M4_EFFECT_TAGS.engine.placementApplied, M10_EFFECT_TAGS.map.placementParityCaptured],
+  requires: [MAP_PROJECTION_EFFECT_TAGS.map.landmassRegionsPlotted],
+  provides: [
+    STANDARD_ENGINE_EFFECT_TAGS.engine.placementApplied,
+    MAP_PROJECTION_EFFECT_TAGS.map.placementParityCaptured,
+  ],
   artifacts: {
     requires: [
       placementArtifacts.placementInputs,
